# api/openapi.yaml
openapi: 3.0.3
info:
  title: TilmanCloud Admin API
  version: 1.0.0
  description: Cloud control plane API
servers:
  - url: http://localhost:8080/
    description: Local development
paths:
...
  /nodes/{nodeId}:
    delete:
      operationId: deleteNode
      summary: Delete a node
      description: Deletes (terminates) a node by its ID
      parameters:
        - name: nodeId
          in: path
          required: true
          description: The ID of the node to delete
          schema:
            type: string
            example: "i-1234567890abcdef0"
      responses:
        '204':
          description: Node deleted successfully
        '404':
          description: Node not found
        '500':
          description: Internal server error
...